requirements.txt
python-telegram-bot==20.6
aiosqlite==0.19.0
bcrypt==4.1.2
httpx==0.25.2
fastapi==0.103.2
uvicorn==0.23.2
